FW: Updated RFG to clearer syntax and added FPGA Timestamp Trigger Logic Unit
- The FPGA Timestamp is now provided by the TLU Module from Tomas @ Desy
- It is slightly modified so that the fpga timestamp counts when enabled
and can be slowed down using the register file divider, the behavior
stays the same
- If the TLU is enabled, the timestamp is reset by the TLU "T0" start of
run, counts in the background, and the value provided to the frame
building logic is updated when a trigger input toggles, synchronised
with the external clock (trigger is set when the calorimeter sees a
particle, updating the timestamp)
- There is a new tlu busy output which is set to 1 after a trigger is
received, which prevents the trigger module to send a new trigger as
long as it stays asserted. The duration of the busy in clock cycles can
be set in the register file. This allows us to "block" timestamp updates
for a while, basically for enough time so that all hits are read from
astropix. A good value will be found by experimenting.