FW/SW: Added SR Config ReadBack
Added CRC Module that can sample one of the lanes SOUT and pack the read
bits in a FIFO - a CRC value is calculated in parallel which can be used
to quickly verify that a config is correct.

Added board driver methods that drives the readback module then also
unpacks the read back bits from the produced FW bytes and returns read
back and expected config bits (from asic config).